+++ web/html/src/components/input/select/Select.tsx
@@ -7,99 +7,9 @@ import AsyncSelect from "react-select/async";
 import { AsyncPaginate as AsyncPaginateSelect } from "react-select-async-paginate";
 
 import withTestAttributes from "./select-test-attributes";
+import { SelectOption, SelectProps } from "./types";
 
-/** Usually options have the shape `{ value: string; label:string }`, but the consumer can define any shape */
-export type Option = { value: string; label: string } | Record<string, unknown>;
-
-type SingleValue<V> = {
-  value?: V;
-
-  onChange?: (newValue: V | undefined) => void;
-
-  /** Set to true to allow multiple selected values */
-  isMulti?: never;
-};
-
-type ValueArray<V> = {
-  value?: V;
-
-  onChange?: (newValue: V) => void;
-
-  /** Set to true to allow multiple selected values */
-  isMulti: true;
-};
-
-type Value<V> = SingleValue<V> | ValueArray<V>;
-
-type CommonSelectProps<T extends Option, V> = {
-  getOptionValue: (option: T) => V;
-  getOptionLabel: (option: T) => V;
-
-  /** Formats option labels in the menu and control as React components */
-  formatOptionLabel?: (option: any, meta: any) => React.ReactNode;
-
-  /** Placeholder for the select value */
-  placeholder?: React.ReactNode;
-
-  /** whether the component's data is loading or not (async) */
-  isLoading?: boolean;
-
-  /** text to display when there are no options to list */
-  emptyText: string | null;
-
-  /** Set to true to allow removing the selected value */
-  isClearable: boolean;
-
-  inputClass?: string;
-
-  /** Id for testing purposes */
-  "data-testid"?: string;
-
-  label?: string;
-
-  name?: string;
-
-  disabled?: boolean;
-};
-
-type SimpleSelectProps<T extends Option, V> = CommonSelectProps<T, V> & {
-  /** Select options */
-  options: T[];
-};
-
-type AsyncSelectProps<T extends Option, V> = CommonSelectProps<T, V> & {
-  /** Default value object if no value is set. This has to be an object corresponding to the rest of the schema. */
-  defaultValueOption?: T;
-
-  paginate?: boolean;
-
-  /**
-   * Function that returns a promise, which is the set of options to be used once the promise resolves.
-   */
-  loadOptions: (searchString: string, callback: (options: T[]) => undefined) => Promise<any> | undefined;
-  cacheOptions?: boolean;
-};
-
-type AsyncPaginateSelectProps<T extends Option, V> = CommonSelectProps<T, V> & {
-  /** Default value object if no value is set. This has to be an object corresponding to the rest of the schema. */
-  defaultValueOption?: T;
-
-  paginate: true;
-  /**
-   * Function that returns a promise with pagination data and a set of options matching the search string
-   * See: https://github.com/vtaits/react-select-async-paginate/tree/master/packages/react-select-async-paginate#loadoptions
-   */
-  loadOptions: (
-    searchString: string,
-    previouslyLoaded: any[],
-    additional?: any
-  ) => Promise<{ options: T[]; hasMore: boolean; additional?: any }>;
-};
-
-type Props<T extends Option, V> = Value<V> &
-  (SimpleSelectProps<T, V> | AsyncSelectProps<T, V> | AsyncPaginateSelectProps<T, V>);
-
-export function Select<T extends Option, V>(props: Props<T, V>) {
+export function Select<T extends SelectOption, V>(props: SelectProps<T, V>) {
   // Make the component controlled. We actually only need this for the async cases, but it's simpler to keep it shared.
   const [value, setValue] = useState(() => {
     // For async, use the default preselected value if available
